+++ src/aux-server/aux-web/aux-projector/scene/BuilderSimulation3D.ts
@@ -52,7 +52,7 @@ export class BuilderSimulation3D extends Simulation3D {
         }
 
         this.simulation.helper.updateBot(this.simulation.helper.userBot, {
-            tags: { 'aux._userChannel': this.simulation.id },
+            tags: { 'aux._user.channel': this.simulation.id },
         });
     }
 }